Committing skeleton chat room up for merge so people tomorrow can work on it
