+++ client/src/components/Chat/ChatMessage.jsx
@@ -5,12 +5,12 @@ import { useCollectionData } from 'react-firebase-hooks/firestore';
 
 const ChatMessage = (props) => {
   const { text, uid, displayName } = props.msg;
-  const messageClass = uid === auth.currentUser.uid ? 'sent' : 'received';
+  const msgType = uid === auth.currentUser.uid ? 'sent' : 'received';
 
   return (
     <>
       <div className='message'>
-        {text}
+        <p>{text}</p>
       </div>
     </>)
 }
